# PERF: GroupBy.apply