```
from asv_bench.benchmarks.groupby import *
self = Apply()
self.setup(4)

%timeit self.time_copy_function_multi_col(4)
451 ms ± 2.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- PR
559 ms ± 1.35 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- master
```