@mzeitlin11 2021-09-15
There's an issue #38358 about the commit that added this `filter` causing a perf regression, would this fix that?